Move entity_source pkg outside of variable_source pkg
