+++ internal/resolution/entity_sources/catalogsource/catalogsource_suite_test.go
