+++ internal/resolution/entity_sources/catalogsource/registry_bundle_to_entity.go
