+++ internal/resolution/entity_sources/catalogsource/registry_grpc_client.go
