+++ internal/resolution/entity_sources/hardcoded/bundle_cache.go
