+++ internal/resolution/variable_sources/olm/olm.go
@@ -5,10 +5,10 @@ import (
 
 	"github.com/operator-framework/deppy/pkg/deppy"
 	"github.com/operator-framework/deppy/pkg/deppy/input"
+	"github.com/operator-framework/operator-controller/internal/resolution/variable_sources/required_package"
 
 	"github.com/operator-framework/operator-controller/internal/resolution/variable_sources/bundles_and_dependencies"
 	"github.com/operator-framework/operator-controller/internal/resolution/variable_sources/crd_constraints"
-	"github.com/operator-framework/operator-controller/internal/resolution/variable_sources/required_package"
 )
 
 var _ input.VariableSource = &OLMVariableSource{}
