+++ internal/resolution/variable_sources/required_package/required_package_test.go
@@ -9,10 +9,10 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/operator-framework/deppy/pkg/deppy"
 	"github.com/operator-framework/deppy/pkg/deppy/input"
+	"github.com/operator-framework/operator-controller/internal/resolution/variable_sources/required_package"
 	"github.com/operator-framework/operator-registry/alpha/property"
 
 	olmentity "github.com/operator-framework/operator-controller/internal/resolution/variable_sources/entity"
-	"github.com/operator-framework/operator-controller/internal/resolution/variable_sources/required_package"
 )
 
 func TestRequiredPackage(t *testing.T) {
